Move request and pool to a separate namespace
